Add ways to include inherited members into generated class docs
This adds a global sphinx option `automodsumm_inherited_members`.
The default is `False` matching the previous behavior.
However, if set to `True` inherited members of classes will be included
into the generated API documentation.

Note that I added no tests for this behavior. There are no tests for the exiting
behavior either and adding tests for those kind of things is a lot of work.
However, I generated docs and checked by hand that it works as exepcted.

closes #109
[skip ci]